requirements/base.txt
flask
flask_jwt_extended
flask-restplus
flask-migrate
flask-sqlalchemy
flask-marshmallow
forgerypy
marshmallow-sqlalchemy
python-dotenv
passlib
psycopg2-binary
